rename to uxi_celery_scheduler
