+++ alembic.ini
@@ -1,6 +1,6 @@
 [alembic]
 # path to migration scripts
-script_location = celery_sqlalchemy_scheduler/db/migrations
+script_location = uxi_celery_scheduler/db/migrations
 
 # template used to generate migration files
 file_template = %%(year)d_%%(month).2d_%%(day).2d_%%(slug)s
