+++ setup.cfg
@@ -33,7 +33,7 @@ ignore_missing_imports = True
 [mypy-setuptools.*]
 ignore_missing_imports = True
 
-[mypy-celery_sqlalchemy_scheduler.db.migrations.*]
+[mypy-uxi_celery_scheduler.db.migrations.*]
 ignore_errors = True
 
 [mypy-alembic.*]
