+++ tests/test_controller.py
@@ -2,9 +2,9 @@ from typing import Dict
 
 from mock import patch
 
-from celery_sqlalchemy_scheduler.controller import schedule_task
-from celery_sqlalchemy_scheduler.data_models import ScheduledTask
-from celery_sqlalchemy_scheduler.db.models import CrontabSchedule, PeriodicTask
+from uxi_celery_scheduler.controller import schedule_task
+from uxi_celery_scheduler.data_models import ScheduledTask
+from uxi_celery_scheduler.db.models import CrontabSchedule, PeriodicTask
 
 
 def test_schedule_task():
